awesomewm: remove theme.layoutlist
It was unused since a long time and caused a weird memory leak.
Because it took me long enough to recognize it as the cause for the
memory leak, I didn't tested further, but I guess it was because I used
a client shape. The internet tells me that may be a problem.