Remove system specific configuration from scripts
Move system specific configurations to a configuration file that gets
loaded by ~/.profile according to the hostname my dotfiles are used on.

That's a first step to enable my dotfiles to be used on different
machines with different hardware setups.

The names of affected scripts are prefixed with "de-" to reflect their
belonging to my desktop environment.